Use one service for updating invoice stock out lines
