+++ server/service/src/invoice_line/outbound_shipment_line/mod.rs
@@ -1,5 +1,3 @@
-pub mod update;
-pub use self::update::*;
 pub mod delete;
 pub use self::delete::*;
