--- server/service/src/invoice_line/prescription_line/update/mod.rs
@@ -1,2 +1,0 @@
-mod update;
-pub use update::*;
