+++ server/service/src/invoice_line/stock_out_line/update/mod.rs
@@ -1,27 +1,68 @@
+use repository::{
+    InvoiceLine, InvoiceLineRow, InvoiceLineRowRepository, RepositoryError, StockLineRow,
+    StockLineRowRepository,
+};
+
 use crate::{
-    invoice_line::{
-        common_update_line::{generate, validate, UpdateInvoiceLine, UpdateInvoiceLineError},
-        query::get_invoice_line,
-    },
+    invoice_line::{query::get_invoice_line, ShipmentTaxUpdate},
     service_provider::ServiceContext,
 };
-use repository::{InvoiceLine, InvoiceLineRowRepository, InvoiceRowType, StockLineRowRepository};
 
-type OutError = UpdateInvoiceLineError;
+mod generate;
+use generate::generate;
+mod validate;
+use validate::validate;
 
-pub fn update_outbound_shipment_line(
+use super::StockOutType;
+
+#[derive(Clone, Debug, PartialEq, Default)]
+pub struct UpdateStockOutLine {
+    pub id: String,
+    pub r#type: Option<StockOutType>,
+    pub item_id: Option<String>,
+    pub stock_line_id: Option<String>,
+    pub number_of_packs: Option<f64>,
+    pub total_before_tax: Option<f64>,
+    pub tax: Option<ShipmentTaxUpdate>,
+    pub note: Option<String>,
+}
+
+#[derive(Clone, Debug, PartialEq)]
+pub enum UpdateStockOutLineError {
+    LineDoesNotExist,
+    DatabaseError(RepositoryError),
+    InvoiceDoesNotExist,
+    InvoiceTypeDoesNotMatch,
+    NoInvoiceType,
+    NotThisStoreInvoice,
+    NotThisInvoiceLine(String),
+    CannotEditFinalised,
+    ItemNotFound,
+    StockLineNotFound,
+    NumberOfPacksBelowOne,
+    ItemDoesNotMatchStockLine,
+    LocationIsOnHold,
+    LocationNotFound,
+    LineDoesNotReferenceStockLine,
+    BatchIsOnHold,
+    UpdatedLineDoesNotExist,
+    StockLineAlreadyExistsInInvoice(String),
+    ReductionBelowZero {
+        stock_line_id: String,
+        line_id: String,
+    },
+}
+
+type OutError = UpdateStockOutLineError;
+
+pub fn update_stock_out_line(
     ctx: &ServiceContext,
-    input: UpdateInvoiceLine,
+    input: UpdateStockOutLine,
 ) -> Result<InvoiceLine, OutError> {
     let updated_line = ctx
         .connection
         .transaction_sync(|connection| {
-            let (line, item, batch_pair, invoice) = validate(
-                &input,
-                &ctx.store_id,
-                &connection,
-                InvoiceRowType::OutboundShipment,
-            )?;
+            let (line, item, batch_pair, invoice) = validate(&input, &ctx.store_id, &connection)?;
 
             let (update_line, batch_pair) = generate(input, line, item, batch_pair, invoice)?;
             InvoiceLineRowRepository::new(&connection).upsert_one(&update_line)?;
@@ -40,14 +81,55 @@ pub fn update_outbound_shipment_line(
     Ok(updated_line)
 }
 
+/// During outbound shipment line / prescription line update, stock line may change thus
+/// validation and updates need to apply to both batches
+pub struct BatchPair {
+    /// Main batch to be updated
+    pub main_batch: StockLineRow,
+    /// Optional previous batch (if batch was changed)
+    pub previous_batch_option: Option<StockLineRow>,
+}
+
+impl BatchPair {
+    /// Calculate reduction amount to apply to main batch
+    pub fn get_main_batch_reduction(
+        &self,
+        input: &UpdateStockOutLine,
+        existing_line: &InvoiceLineRow,
+    ) -> f64 {
+        // Previous batch exists, this mean new batch was requested means:
+        // - reduction should be number of packs from input (or existing line if number of pack is missing in input)
+        if self.previous_batch_option.is_some() {
+            input
+                .number_of_packs
+                .unwrap_or(existing_line.number_of_packs)
+        } else {
+            // Previous batch does not exists, this mean updating existing batch, thus:
+            // - reduction is the difference between input and existing line number of packs
+            if let Some(number_of_packs) = &input.number_of_packs {
+                *number_of_packs - existing_line.number_of_packs
+            } else {
+                // No changes in input, no reduction
+                0.0
+            }
+        }
+    }
+}
+
+impl From<RepositoryError> for UpdateStockOutLineError {
+    fn from(error: RepositoryError) -> Self {
+        UpdateStockOutLineError::DatabaseError(error)
+    }
+}
+
 #[cfg(test)]
 mod test {
     use repository::{
         mock::{
-            mock_inbound_shipment_a_invoice_lines, mock_item_b, mock_item_b_lines,
-            mock_outbound_shipment_a_invoice_lines, mock_outbound_shipment_b_invoice_lines,
-            mock_outbound_shipment_c, mock_outbound_shipment_c_invoice_lines,
-            mock_outbound_shipment_no_stock_line, mock_stock_line_a, mock_stock_line_b,
+            mock_item_b, mock_item_b_lines, mock_outbound_shipment_a_invoice_lines,
+            mock_outbound_shipment_b_invoice_lines, mock_outbound_shipment_c,
+            mock_outbound_shipment_c_invoice_lines, mock_outbound_shipment_no_stock_line,
+            mock_prescription_a_invoice_lines, mock_stock_line_a, mock_stock_line_b,
             mock_stock_line_location_is_on_hold, mock_stock_line_on_hold, mock_store_a,
             mock_store_b, mock_store_c, MockDataInserts,
         },
@@ -58,19 +140,16 @@ mod test {
 
     use crate::{
         invoice::outbound_shipment::{UpdateOutboundShipment, UpdateOutboundShipmentStatus},
-        invoice_line::common_update_line::{
-            UpdateInvoiceLine, UpdateInvoiceLineError as ServiceError,
+        invoice_line::stock_out_line::{
+            StockOutType, UpdateStockOutLine, UpdateStockOutLineError as ServiceError,
         },
         service_provider::ServiceProvider,
     };
 
     #[actix_rt::test]
-    async fn update_outbound_shipment_line_errors() {
-        let (_, _, connection_manager, _) = setup_all(
-            "update_outbound_shipment_line_errors",
-            MockDataInserts::all(),
-        )
-        .await;
+    async fn update_stock_out_line_errors() {
+        let (_, _, connection_manager, _) =
+            setup_all("update_stock_out_line_errors", MockDataInserts::all()).await;
 
         let service_provider = ServiceProvider::new(connection_manager, "app_data");
         let mut context = service_provider
@@ -80,10 +159,11 @@ mod test {
 
         // LineDoesNotExist
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = "invalid".to_string();
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::LineDoesNotExist)
@@ -91,34 +171,38 @@ mod test {
 
         // NotThisStoreInvoice
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(10.0);
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::NotThisStoreInvoice)
         );
 
-        // NotAnOutboundShipment
+        // InvoiceTypeDoesNotMatch
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
-                    r.id = mock_inbound_shipment_a_invoice_lines()[0].id.clone();
+                inline_init(|r: &mut UpdateStockOutLine| {
+                    r.id = mock_prescription_a_invoice_lines()[0].id.clone();
+                    r.number_of_packs = Some(10.0);
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
-            Err(ServiceError::NotAnOutboundShipment)
+            Err(ServiceError::InvoiceTypeDoesNotMatch)
         );
 
         context.store_id = mock_store_c().id;
         // CannotEditFinalised
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_b_invoice_lines()[0].id.clone();
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::CannotEditFinalised)
@@ -126,10 +210,11 @@ mod test {
 
         // LineDoesNotReferenceStockLine
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_no_stock_line()[0].id.clone();
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::LineDoesNotReferenceStockLine)
@@ -138,11 +223,12 @@ mod test {
         context.store_id = mock_store_b().id;
         // ItemNotFound
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.item_id = Some("invalid".to_string());
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::ItemNotFound)
@@ -150,11 +236,12 @@ mod test {
 
         // StockLineNotFound
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.stock_line_id = Some("invalid".to_string());
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::StockLineNotFound)
@@ -162,11 +249,12 @@ mod test {
 
         // NumberOfPacksBelowOne
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(0.0);
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::NumberOfPacksBelowOne)
@@ -174,12 +262,13 @@ mod test {
 
         // ItemDoesNotMatchStockLine
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.item_id = Some(mock_item_b().id.clone());
                     r.stock_line_id = Some(mock_stock_line_a().id.clone());
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::ItemDoesNotMatchStockLine)
@@ -187,12 +276,13 @@ mod test {
 
         // LocationIsOnHold
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.item_id = Some(mock_stock_line_location_is_on_hold()[0].item_id.clone());
                     r.stock_line_id = Some(mock_stock_line_location_is_on_hold()[0].id.clone());
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::LocationIsOnHold)
@@ -200,12 +290,13 @@ mod test {
 
         // BatchIsOnHold
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.stock_line_id = Some(mock_stock_line_on_hold()[0].id.clone());
                     r.item_id = Some(mock_stock_line_on_hold()[0].item_id.clone());
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::BatchIsOnHold)
@@ -213,11 +304,12 @@ mod test {
 
         // ReductionBelowZero
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(100.0);
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::ReductionBelowZero {
@@ -231,28 +323,24 @@ mod test {
 
         // StockLineAlreadyExistsInInvoice
         assert_eq!(
-            service.update_outbound_shipment_line(
+            service.update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.stock_line_id = Some(mock_item_b_lines()[0].id.clone());
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             ),
             Err(ServiceError::StockLineAlreadyExistsInInvoice(
                 mock_outbound_shipment_a_invoice_lines()[1].id.clone()
             ))
         );
-
-        //TODO InvoiceDoesNotExist
     }
 
     #[actix_rt::test]
-    async fn update_outbound_shipment_line_success() {
-        let (_, connection, connection_manager, _) = setup_all(
-            "update_outbound_shipment_line_success",
-            MockDataInserts::all(),
-        )
-        .await;
+    async fn update_stock_out_line_success() {
+        let (_, connection, connection_manager, _) =
+            setup_all("update_stock_out_line_success", MockDataInserts::all()).await;
 
         // helpers to compare total
         let stock_line_for_invoice_line = |invoice_line: &InvoiceLineRow| {
@@ -286,12 +374,13 @@ mod test {
             .unwrap();
 
         service
-            .update_outbound_shipment_line(
+            .update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_c_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(2.0);
                     r.total_before_tax = Some(18.00);
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             )
             .unwrap();
@@ -346,13 +435,14 @@ mod test {
             )
             .unwrap();
         service
-            .update_outbound_shipment_line(
+            .update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_c_invoice_lines()[0].id.clone();
                     r.stock_line_id = Some(mock_stock_line_b().id.clone());
                     r.number_of_packs = Some(2.0);
                     r.total_before_tax = Some(10.99);
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             )
             .unwrap();
@@ -388,12 +478,13 @@ mod test {
 
         context.store_id = mock_store_b().id;
         service
-            .update_outbound_shipment_line(
+            .update_stock_out_line(
                 &context,
-                inline_init(|r: &mut UpdateInvoiceLine| {
+                inline_init(|r: &mut UpdateStockOutLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(15.0);
                     r.total_before_tax = Some(10.99);
+                    r.r#type = Some(StockOutType::OutboundShipment);
                 }),
             )
             .unwrap();
